+++ crates/store/re_datafusion/src/dataframe_query_common.rs
@@ -23,8 +23,8 @@ use re_dataframe::external::re_chunk_store::ChunkStore;
 use re_dataframe::{Index, QueryExpression};
 use re_log_encoding::codec::wire::decoder::Decode as _;
 use re_log_types::EntryId;
-use re_protos::cloud::v1alpha1::DATASET_MANIFEST_ID_FIELD_NAME;
 use re_protos::cloud::v1alpha1::ext::{Query, QueryLatestAt, QueryRange};
+use re_protos::cloud::v1alpha1::{DATASET_MANIFEST_ID_FIELD_NAME, QueryDatasetResponse};
 use re_protos::cloud::v1alpha1::{GetDatasetSchemaRequest, QueryDatasetRequest};
 use re_protos::common::v1alpha1::ext::ScanParameters;
 use re_protos::headers::RerunHeadersInjectorExt as _;
@@ -106,10 +106,10 @@ impl DataframeQueryTableProvider {
         let query = query_from_query_expression(query_expression);
 
         let fields_of_interest = [
-            "chunk_partition_id",
-            "chunk_id",
-            "rerun_partition_layer",
-            "chunk_key",
+            QueryDatasetResponse::PARTITION_ID,
+            QueryDatasetResponse::CHUNK_ID,
+            QueryDatasetResponse::PARTITION_LAYER,
+            QueryDatasetResponse::CHUNK_KEY,
         ]
         .into_iter()
         .map(String::from)
